Add AppLicense functionality.
AppLicense allows operators to avoid spawning too many jobs for
the number of total real-world (e.g. flexlm) licenses available to run
them.

1) Create a variable called license/<licensename> with a field "count"
   equal to the number of licenses you want simultaneously available.
2) Add one or more app_license blocks to your job, like
  app_license {
    license = "licensename"
    count = 1 # how many licenses this job requires to run of that type
  }

At plan time, it will calculate the number of licenses available and if
there are enough, it will allow the job to launch. Otherwise, it will
send the job back for rescheduling.